Add main.cpp and makefile c++/python tests
